Show/Hide Python 3 ready packages
Now that the scales have tipped way in favor of Python 3, make it
easy to see ones that still need porting.

CSS button rounding moved into code since last-child/first-child
pseudo selectors are absolute and not "last/first visible"

Not married to button size and/or location.  Changes as necessary.